pyproject.toml
[tool.poetry]
name = "ui"
version = "0.0.0"
description = "Beautifully designed UI components that you can copy and paste into your PyTempl apps."
authors = ["Saurabh Ghanekar <ghanekarsaurabh8@gmail.com>"]
license = "BSD-2-Clause"
readme = "README.md"
homepage = "https://github.com/pytempl/ui"
repository = "https://github.com/pytempl/ui"
documentation = "https://github.com/pytempl/ui"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]

[tool.poetry.dependencies]
python = "^3.10"
pytempl = "^0"
pytempl-icon = "^0"

[tool.poetry.group.dev.dependencies]
rich = "^13.7"
black = "^24.8"
ruff = "^0.5"
pre-commit = "^3.8"
tomli = "^2.0"

[tool.black]
line-length = 88


[tool.ruff]
line-length = 88


[tool.ruff.lint]
select = [
    # "D",      # pydocstyle
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG001", # unused arguments in functions
]
ignore = [
    "E501", # line too long, handled by black
    "B008", # do not perform function calls in argument defaults
    "W191", # indentation contains tabs
    "B904", # Allow raising exceptions without from e, for HTTPException
]


[tool.ruff.lint.pydocstyle]
convention = "google"


[tool.versioning.files]
version_variable = ["src/ui/__init__.py:__version__"]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
